refactor: extract service
Extract service for searching packuments on a registry.
